+++ src/solutions/p10.rs
@@ -1,6 +1,6 @@
 use maus_project_euler::prime_sieve_hashset;
 
-pub fn s_v1(n: &u32) -> u64 {
+pub fn s_v1(n: &u64) -> u64 {
     let primes = prime_sieve_hashset(n);
 
     let mut sum: u64 = 0;
