+++ src/solutions/p5.rs
@@ -1,4 +1,4 @@
-pub fn s_v1() -> u32 {
+pub fn s_v1() -> u64 {
     let mut i = 2;
     while !is_divisible(&i) {
         i += 1;
@@ -6,7 +6,7 @@ pub fn s_v1() -> u32 {
     i
 }
 
-fn is_divisible(n: &u32) -> bool {
+fn is_divisible(n: &u64) -> bool {
     for i in [20, 19, 18, 17, 16, 15, 14, 13, 11] {
         if n % i != 0 {
             return false
